+++ changelog.md
@@ -3,7 +3,8 @@
 ## Next
 
 * Updated dependencies.
-* Remove `Promise` polyfill. For optimal bundle size consumers can handle it.
+* Significantly reduced the bundle size to < 4 KB by simplifying Babel helpers and reusing the [`object-assign`](https://npm.im/object-assign) React dependency.
+* Removed the `Promise` polyfill; consumers can polyfill as required for optimal bundle size.
 * Document required polyfills in the readme.
 * Updated ESLint config:
   * `parserOptions` is unnecessary when using `babel-eslint`.